refactor: rename plugin-education package to education
The plugin-education package is a shared workspace library consumed by
plugins (eol, etc.) for first-run educational modals, not a registered
plugin itself. The plugin- prefix has been a recurring source of
confusion — there is no /api/plugin-education/* route, no
plugin-education tables, and no entry in lib/plugins/repository.ts.

This change drops the misleading prefix everywhere:

- Directory: ctf/packages/plugin-education/ → ctf/packages/education/
- npm name:  @ctf/plugin-education          → @ctf/education
- Inventory: plugin-education-feature-inventory.md
             → education-feature-inventory.md

Updates all consumers (eol package.json + education.tsx imports +
README), both pnpm-workspace.yaml files, both pnpm-lock.yaml files,
AGENTS.md monorepo layout, and the EOL inventory's "Educational Module
Integration" reference. Logs the rename in the education inventory's
Change Log so the package's history stays traceable.

Verified clean workspace typecheck (web + mobile + shared) and clean
web/android parity check after the rename.
