+++ ctf/packages/education/components/EducationModal.tsx
