+++ ctf/packages/education/content/eol.md
