+++ ctf/packages/education/content/loader.ts
