+++ ctf/packages/education/feature-inventory.md
@@ -1,0 +1,3 @@
+# Education Module Feature Inventory
+
+See the canonical inventory at [ctf/docs/developer/ctf-plugin-feature-inventories/education-feature-inventory.md](../../docs/developer/ctf-plugin-feature-inventories/education-feature-inventory.md)
