+++ ctf/packages/education/index.ts
