+++ ctf/packages/education/schema.ts
