+++ ctf/packages/education/tsconfig.json
