+++ ctf/packages/eol/package.json
@@ -5,7 +5,7 @@
   "types": "index.tsx",
   "private": true,
   "dependencies": {
-    "@ctf/plugin-education": "workspace:*",
+    "@ctf/education": "workspace:*",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
     "react": "19.2.5",
