--- ctf/packages/plugin-education/feature-inventory.md
@@ -1,3 +1,0 @@
-# Plugin Education Module Feature Inventory
-
-See the canonical inventory at [ctf/docs/developer/ctf-plugin-feature-inventories/plugin-education-feature-inventory.md](../../docs/developer/ctf-plugin-feature-inventories/plugin-education-feature-inventory.md)
